requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.4
attrs==21.2.0
beautifulsoup4==4.9.3
black==21.5b1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.2.3
django-allauth==0.44.0
django-cleanup==5.2.0
django-debug-toolbar==3.1
django-environ==0.4.5
django-extensions==3.1.3
django-taggit==1.4.0
django-widget-tweaks==1.4.8
filelock==3.0.12
flake8==3.9.2
flake8-import-order==0.18.1
idna==2.10
iniconfig==1.1.1
isort==5.8.0
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.9
pathspec==0.8.1
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.3.0
python3-openid==3.2.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
tox==3.23.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.4
virtualenv==20.4.6
Werkzeug==2.0.1
